+++ R/preprocessing_feature_selection.R
@@ -14,9 +14,9 @@
 #' @param feature_selection_method A string value indication the feature selection method.
 #' The imputation method must be one of 'VI', 'MCFS', 'MI', or 'BORUTA'.
 #' @param max_features A positive integer value describing the desired number of
-#' selected features. Initial value set as 'default' which is 10 for `VI` and `MI`, and
-#' NULL (number of relevant features chosen by the method) for `MCFS`.
-#' Only `MCFS` can use the NULL value. `BORUTA` doesn't use this parameter.
+#' selected features. Initial value set as 'default' which is min(10, ncol(data) - 1)
+#' for `VI` and `MI`, and NULL (number of relevant features chosen by the method)
+#' for `MCFS`. Only `MCFS` can use the NULL value. `BORUTA` doesn't use this parameter.
 #' @param nperm An integer describing the number of permutations performed, relevant
 #' for the `VI` method. By default set to 1.
 #' @param cutoffPermutations An non-negative integer value that determines the number of permutation
@@ -50,7 +50,7 @@ preprocessing_feature_selection <- function(data,
 
   if (feature_selection_method == 'VI') {
     if (max_features == 'default') {
-      max_features <- 10
+      max_features <- min(10, ncol(data) - 1)
     }
     fs_data <- select_vi(data, y, nperm = 1, max_features = max_features)
   } else if (feature_selection_method == 'MCFS') {
@@ -60,7 +60,7 @@ preprocessing_feature_selection <- function(data,
     fs_data <- select_mcfs(data, y, cutoffPermutations = cutoffPermutations, threadsNumber = threadsNumber, max_features = max_features)
   } else if (feature_selection_method == 'MI') {
     if (max_features == 'default') {
-      max_features <- 10
+      max_features <- min(10, ncol(data) - 1)
     }
     fs_data <- select_mi_varrank(data, y, max_features = max_features)
   } else if (feature_selection_method == 'BORUTA') {