+++ R/preprocessing_imputation.R
@@ -17,9 +17,9 @@
 #' @param imputation_method A string value indication the imputation method. The
 #' imputation method must be one of 'median-other', 'median-frequency', 'knn', or 'mice'.
 #' @param k An integer describing the number of nearest neighbours to use. By default
-#' set to 10. The parameter applicable only if selection `imputation_method` is 'knn'.
+#' set to 10. The parameter applicable only if selected `imputation_method` is 'knn'.
 #' @param m An integer describing the number of multiple imputations to use.
-#' By default set to 5. The parameter applicable only if selection
+#' By default set to 5. The parameter applicable only if selected
 #' `imputation_method` is 'mice'.
 #'
 #' @return Imputed dataset.